Reprice the sub-$1 tiers: gpt-reserve and luna 1 → 0.2
User ruling 2026-09-03, the follow-on to fractional credits: "i think if we
are supporting fractional credts we should reprice agents that are under
$1/m". Fractional seats landed (d2677f1) re-pricing nothing; this is the
repricing.

WHICH TIERS QUALIFY. Read off the repo's own sourced price tables
(providers.CODEX_PRICES, providers.ANTIGRAVITY_PRICES) rather than inferred
from the seat numbers. Standing $/M input: fable ≫1, opus 5, sonnet 2,
haiku 1.00, sol 5.00 ($4 promo), terra 2.00, gpt-reserve 0.20, luna 0.20,
flash 1.50 ($0.75 launch price), pro 2.00. Only gpt-reserve and luna are
under $1, and both land at 0.2 — twice the 0.10 floor, so nothing is
clamped. Everything else keeps the >=$1 branch's floor, haiku included:
exactly $1.00 is not below $1.

flash stays 1 (user, directly: "keep flash on 1 for now"). Its $0.75 is
launch pricing through 2026-12-31 and a promo never sets a seat — the
sonnet-intro precedent, re-affirmed for sol 2026-08-28 and for flash by the
user himself 2026-09-02. Pricing it at 0.75 would also force the seat back
UP on 2027-01-01, and up is the one direction that can strand a saved org.

IT NEEDS A MIGRATION, AND THE ZERO-MIGRATION PROPERTY SURVIVES ANYWAY.
Org.create snapshots TIERS into the doc and the load merge is add-only, so
editing the constant reaches NO existing org — measured: all three live docs
on this machine carried `gpt-reserve: 1, luna: 1`, i.e. without a migration
block this ships to nobody. Added on the sonnet 3→2 precedent, migrating
ONLY the old shipped default of 1. But no credit quantity is rewritten: a
node records `model` and `grant`, never a seat, and seat_cost reads the doc's
tier table on every call, so committed/free re-derive the instant the number
lands. It is a dict-value edit, not a data migration.

IT MOSTLY LOOSENS — AND ONE THING TIGHTENS. committed falls, free rises,
free() >= 0 cannot start failing. But `_check_tier_ceiling` compares seats as
an ORDERING, and tiers that tied at 1 no longer tie: a `max_tier="luna"`
ceiling used to admit haiku at five times luna's price because 1 > 1 is
false, and now refuses it. That is the tie-collapse bug in another costume,
so the new behaviour is the correct one — but it IS a change to a saved
ceiling and it is pinned by test rather than left to be found. Measured: no
live org had a kiosk ceiling set at all.

FIVE PRICE TABLES, NOT ONE. The seat is copied into providers.py's derived
views, OrgCanvas's `tree.tiers ?? {…}` fallback, and shared.ts's TIER_SEAT /
CODEX_TIER_SEAT / ANTIGRAVITY_TIER_SEAT (which `anyTierSeat` prices the desk,
the cards and the hire chips through). chiptips §8 guarded the OrgCanvas copy
and caught it; nothing guarded shared.ts, which sat at the old 1 until a
rehire panel rendered "seat 1". New §8b parses all three shared.ts tables out
of source and compares them to ledger.TIERS — verified anti-vacuous by
reverting the table and watching it fail.

Also fixed, all consequences of the same fractional change already on main:
- providers.py's CODEX_TIERS / ANTIGRAVITY_TIERS / CLAUDE_TIERS were declared
  dict[str, int] while deriving from a dict[str, float] — three live pyright
  reportAssignmentType errors — and TierInfo.seat was int. Harmless while
  every static seat was whole; wrong the moment one is 0.2.
- test_ledger_authority's randomized conservation property asserted every
  grant `isinstance(int)`. Stale, not violated: on main today a switch onto a
  $0.20 OpenRouter favorite already yields a float grant, and the check only
  survived because the pool it draws from was all-whole. Replaced with the
  invariant that actually holds and catches more — a non-bool number exactly
  on the 0.01 credit grid — and the shadow derivation quantised the way the
  ledger quantises its own (a naive float sum produced 1.2000000000000028
  against a correct free() of 1.2).
- chiptips' `\(-\d+\)` badge pattern and its `parts()` cost parser were
  integer-only: "(-0.2)" did not mis-parse, it failed to match, which reads
  as "no cost badge at all".
- Six "floored to 1" claims (mcptool ×2, openrouter, adding-a-provider,
  configuration, setup-guide) have been false since fractional seats landed.

NOT DONE, DELIBERATELY: the four orphaned `or-` tier rows in the live org docs
(or-z-ai-glm-5-3-flash 1, or-deepseek-deepseek-v4-flash-latest 1,
or-moonshotai-kimi-k3 3, or-x-ai-grok-4-6 2). A favorite's seat is snapshotted
at add time and deselecting keeps the org-doc row, so two probably-sub-$1
models are frozen at 1 — and because the merge is add-only, even re-favouriting
them would not repair the row. Repricing them needs a live catalog lookup,
which a load hook must never do. Measured: no node, live or archived, runs on
any `or-` tier. Named rather than fixed; DECISIONS entry follows.

Tests: backend test_providers 35/35 (11 new), test_ledger_authority 194/194
(12 new), frontend 466/466 (1 new). Docs and README moved with the code.

+++ backend/orgtree/providers.py
@@ -54,7 +54,11 @@ class TierInfo(TypedDict):
     """One tier as the UI needs it — name, price band, default model id."""
     tier: str
     provider: str
-    seat: int
+    #: FLOAT, not int (2026-09-03): a sub-$1 tier costs a fraction of a
+    #: credit — gpt-reserve and luna are 0.2 — and the OpenRouter favorites
+    #: this shape is shared with have been fractional since `seat_for` grew
+    #: its below-$1 branch. The frontend already types it `number`.
+    seat: float
     model: str
     letter: str
 
@@ -73,9 +77,12 @@ _CODEX_LETTER: Final[dict[str, str]] = {
 #: views derive from them so there is exactly one copy to drift. Seat rule
 #: (user ruling 2026-08-28, ask card): STANDING API $ per M input — sol $5
 #: standard (the $4 promo, through ≥2026-11-21, never sets a seat), terra
-#: $2, and gpt-reserve/luna $0.20 floored to 1.
+#: $2, and gpt-reserve/luna $0.20 → 0.2 each since the sub-$1 repricing
+#: (user ruling 2026-09-03); they used to floor to 1, which made the four
+#: bands read 1·1·2·5 and lost luna's 10× advantage over terra.
 _CODEX_TIER_NAMES: Final = ("gpt-reserve", "luna", "terra", "sol")
-CODEX_TIERS: Final[dict[str, int]] = {
+#: float, not int: this is THE table the codex fractions live in.
+CODEX_TIERS: Final[dict[str, float]] = {
     t: _LEDGER_TIERS[t] for t in _CODEX_TIER_NAMES}
 CODEX_MODELS: Final[dict[str, str]] = {
     t: _LEDGER_MODELS[t] for t in _CODEX_TIER_NAMES}
@@ -117,7 +124,10 @@ _ANTIGRAVITY_LETTER: Final[dict[str, str]] = {"flash": "F", "pro": "P"}
 #: seat), flash $1.50 standing (3.8-flash's $0.75 is launch pricing through
 #: 2026-12-31, and a promo never sets a seat) → 1.
 _ANTIGRAVITY_TIER_NAMES: Final = ("flash", "pro")
-ANTIGRAVITY_TIERS: Final[dict[str, int]] = {
+#: float for the same reason CODEX_TIERS is, even though neither antigravity
+#: seat is fractional today: the type follows ledger.TIERS, not the values
+#: that happen to be in it.
+ANTIGRAVITY_TIERS: Final[dict[str, float]] = {
     t: _LEDGER_TIERS[t] for t in _ANTIGRAVITY_TIER_NAMES}
 ANTIGRAVITY_MODELS: Final[dict[str, str]] = {
     t: _LEDGER_MODELS[t] for t in _ANTIGRAVITY_TIER_NAMES}
@@ -129,7 +139,7 @@ ANTIGRAVITY_MODELS: Final[dict[str, str]] = {
 #: cannot: it answers "claude" for both, deliberately (see its docstring). A
 #: gate keyed on `provider_of` alone would refuse a typo'd tier with a message
 #: about installing Claude Code.
-CLAUDE_TIERS: Final[dict[str, int]] = {
+CLAUDE_TIERS: Final[dict[str, float]] = {
     t: seat for t, seat in _LEDGER_TIERS.items()
     if t not in _CODEX_TIER_NAMES and t not in _ANTIGRAVITY_TIER_NAMES}
